Refactor VmrManager into smaller classes, improve DI (#1987)
This commit does not contain almost any functional changes, only prepares the code for a follow-up change (https://github.com/dotnet/arcade/issues/10261) to make the changes smaller.

List of changes:
- `VmrManager` broken down into `VmrInitializer` and `VmrUpdater`
- Added better DI registration so that it supports both CLI and MSBuild environments
- `ProcessManager` now accepts `CancellationToken`
- Patch creation is now cancellable
- Removed the `ignoreWorkingTree` option as it won't work together with source build patches
