+++ src/Microsoft.DotNet.Darc/src/DarcLib/VirtualMonoRepo/VmrRegistrations.cs
@@ -1,0 +1,44 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System;
+using Microsoft.DotNet.DarcLib.Helpers;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.DependencyInjection.Extensions;
+
+#nullable enable
+namespace Microsoft.DotNet.DarcLib.VirtualMonoRepo;
+
+public static class VmrRegistrations
+{
+    public static IServiceCollection AddVmrManagers(
+        this IServiceCollection services,
+        string gitLocation,
+        string vmrPath,
+        string tmpPath)
+    {
+        RegisterManagers(services, gitLocation);
+        services.TryAddSingleton<IVmrManagerConfiguration>(new VmrManagerConfiguration(vmrPath, tmpPath));
+        return services;
+    }
+
+    public static IServiceCollection AddVmrManagers(
+        this IServiceCollection services,
+        string gitLocation,
+        Func<IServiceProvider, IVmrManagerConfiguration> configure)
+    {
+        RegisterManagers(services, gitLocation);
+        services.TryAddSingleton<IVmrManagerConfiguration>(configure);
+        return services;
+    }
+
+    private static void RegisterManagers(IServiceCollection services, string gitLocation)
+    {
+        services.TryAddTransient<IProcessManager>(sp => ActivatorUtilities.CreateInstance<ProcessManager>(sp, gitLocation));
+        services.TryAddTransient<ISourceMappingParser, SourceMappingParser>();
+        services.TryAddTransient<IVmrManagerFactory, VmrManagerFactory>();
+        services.TryAddTransient<IVmrUpdater>(sp => sp.GetRequiredService<IVmrManagerFactory>().CreateVmrUpdater().GetAwaiter().GetResult());
+        services.TryAddTransient<IVmrInitializer>(sp => sp.GetRequiredService<IVmrManagerFactory>().CreateVmrInitializer().GetAwaiter().GetResult());
+    }
+}
